SUbiendo ejercicios hechos hasta ahora. Commit inicial

+++ src/ProgramacionModular/ejer5.java
@@ -13,7 +13,7 @@ public class ejer5 {
     public static void Tablamultiplicar(double num){
         double result;
         result = 0;
-        for (double i = 0; i <= 10; i++){
+        for (int i = 0; i <= 10; i++){
             result = i * num;
             System.out.println(result);
